CP - Yeri

+++ Bendiciones/frmBuscarClaseParticular.cs
@@ -28,30 +28,39 @@ namespace Bendiciones
 
         private void btnSeleccionar_Click(object sender, EventArgs e)
         {
-            claseSeleccionada = clases[dgvClasesParticulares.CurrentRow.Index];
-            this.DialogResult = DialogResult.OK;
+            if(cliente == null)
+            {
+                MessageBox.Show("Debe seleccionar primero a un cliente", "Mensaje de advertencia", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                
+            }
+            else if (dgvClasesParticulares.RowCount == 0)
+            {
+                MessageBox.Show("El cliente no tiene clases particulares", "Mensaje de advertencia", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+            }
+            else
+            {
+                claseSeleccionada = clases[dgvClasesParticulares.CurrentRow.Index];
+                this.DialogResult = DialogResult.OK;
+            }
+            
         }
 
 
         private void btnEliminar_Click(object sender, EventArgs e)
         {
-            claseSeleccionada = (Service.claseParticular)dgvClasesParticulares.CurrentRow.DataBoundItem;
-            Program.dbController.eliminarClaseParticular(claseSeleccionada.id_servicio);
-            frmMensaje mensaje = new frmMensaje("Se ha eliminado correctamente", "Mensaje Confirmacion", "Confirmar");
-            this.Close();
-        }
-
-        private void btnBuscar_Click(object sender, EventArgs e)
-        {
-            //si no es vacío
-            foreach ( Service.claseParticular c in Program.dbController.listarClasesParticularesPorCliente(Cliente.idPersona))
+            if (cliente == null || dgvClasesParticulares.RowCount == 0)
             {
-                clases.Add(c);
-                Object[] filaClase = new Object[3];
-                filaClase[0] = c.nombre;
-                filaClase[1] = c.precio;
-                filaClase[2] = c.fecha;
-                dgvClasesParticulares.Rows.Add(filaClase);
+                MessageBox.Show("No hay clase particular seleccionada", "Mensaje de advertencia", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+            } else
+            {
+                claseSeleccionada = clases[dgvClasesParticulares.CurrentRow.Index];
+                Console.WriteLine("id clase particular:");
+                Console.WriteLine(claseSeleccionada.id_servicio);
+                Program.dbController.eliminarClaseParticular(claseSeleccionada.id_servicio);
+                MessageBox.Show("Se ha eliminado correctamente", "Mensaje Confirmacion", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                clases.Remove(ClaseSeleccionada);
+                dgvClasesParticulares.Rows.Remove(dgvClasesParticulares.CurrentRow);
+                this.Close();
             }
             
         }
@@ -65,6 +74,22 @@ namespace Bendiciones
                 Cliente = formBuscarCliente.ClienteSeleccionado;
                 txtDNI.Text = Cliente.dni;
                 txtNombreCliente.Text = Cliente.nombre;
+                dgvClasesParticulares.RowCount = 0;
+                //lista las sus clases particulares
+                IEnumerable<Service.claseParticular> cp = Program.dbController.listarClasesParticularesPorCliente(Cliente.idPersona);
+                if(cp != null)
+                {
+                    foreach (Service.claseParticular c in Program.dbController.listarClasesParticularesPorCliente(Cliente.idPersona))
+                    {
+                        clases.Add(c);
+                        Object[] filaClase = new Object[3];
+                        filaClase[0] = c.nombre;
+                        filaClase[1] = c.precio;
+                        filaClase[2] = c.fecha.ToShortDateString();
+                        dgvClasesParticulares.Rows.Add(filaClase);
+                    }
+                }
+                
             }
         }
